+++ src/addons/display_commander/hooks/streamline_hooks.hpp
@@ -3,7 +3,6 @@
 // Source Code <Display Commander>
 
 #include <cstdint>
-#include <vector>
 
 #include <windows.h>